# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CourtFlow API",
    "version": "1.0.0"
  },
  "paths": {
    "/health": {
      "get": {
        "summary": "Health",
        "operationId": "health_health_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      }
    },
    "/courts": {
      "get": {
        "summary": "List Courts",
        "operationId": "list_courts_courts_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create Court",
        "operationId": "create_court_courts_post",
        "parameters": [
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CourtInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courts/availability": {
      "get": {
        "summary": "Availability",
        "operationId": "availability_courts_availability_get",
        "parameters": [
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Date"
            }
          },
          {
            "name": "time",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
              "title": "Time"
            }
          },
          {
            "name": "duration",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "multipleOf": 30,
              "maximum": 240,
              "minimum": 30,
              "default": 60,
              "title": "Duration"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courts/{court_id}": {
      "get": {
        "summary": "Court Detail",
        "operationId": "court_detail_courts__court_id__get",
        "parameters": [
          {
            "name": "court_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Court Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/courts/{court_id}/schedules": {
      "post": {
        "summary": "Set Schedule",
        "operationId": "set_schedule_courts__court_id__schedules_post",
        "parameters": [
          {
            "name": "court_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Court Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create User",
        "operationId": "create_user_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserInput"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rentals": {
      "post": {
        "summary": "Create Rental",
        "operationId": "create_rental_rentals_post",
        "parameters": [
          {
            "name": "x-user-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RentalInput"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Admin History",
        "operationId": "admin_history_rentals_get",
        "parameters": [
          {
            "name": "court_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Court Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}/rentals": {
      "get": {
        "summary": "User History",
        "operationId": "user_history_users__user_id__rentals_get",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "User Id"
            }
          },
          {
            "name": "x-user-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rentals/{rental_id}/confirm": {
      "post": {
        "summary": "Confirm",
        "operationId": "confirm_rentals__rental_id__confirm_post",
        "parameters": [
          {
            "name": "rental_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rental Id"
            }
          },
          {
            "name": "x-admin-token",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-Admin-Token"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmationInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/rentals/{rental_id}/cancel": {
      "post": {
        "summary": "Cancel",
        "operationId": "cancel_rentals__rental_id__cancel_post",
        "parameters": [
          {
            "name": "rental_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Rental Id"
            }
          },
          {
            "name": "x-user-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "X-User-Key"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ConfirmationInput": {
        "properties": {
          "payment": {
            "type": "string",
            "enum": [
              "Pendiente",
              "Pagado"
            ],
            "title": "Payment",
            "default": "Pendiente"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "title": "ConfirmationInput"
      },
      "CourtInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 70,
            "minLength": 3,
            "title": "Name"
          },
          "venue": {
            "type": "string",
            "maxLength": 70,
            "minLength": 3,
            "title": "Venue"
          },
          "sport": {
            "type": "string",
            "enum": [
              "Fútbol",
              "Vóley",
              "Básquet"
            ],
            "title": "Sport"
          },
          "capacity": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 2.0,
            "title": "Capacity"
          },
          "hourly_rate": {
            "type": "integer",
            "maximum": 100000.0,
            "minimum": 100.0,
            "title": "Hourly Rate"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "venue",
          "sport",
          "capacity",
          "hourly_rate"
        ],
        "title": "CourtInput"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "RentalInput": {
        "properties": {
          "court_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Court Id"
          },
          "user_id": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "User Id"
          },
          "date": {
            "type": "string",
            "format": "date",
            "title": "Date"
          },
          "time": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Time"
          },
          "duration": {
            "type": "integer",
            "multipleOf": 30.0,
            "maximum": 240.0,
            "minimum": 30.0,
            "title": "Duration"
          },
          "event": {
            "type": "string",
            "maxLength": 120,
            "minLength": 3,
            "title": "Event"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "court_id",
          "user_id",
          "date",
          "time",
          "duration",
          "event"
        ],
        "title": "RentalInput"
      },
      "ScheduleInput": {
        "properties": {
          "weekday": {
            "type": "integer",
            "maximum": 6.0,
            "minimum": 0.0,
            "title": "Weekday"
          },
          "opens": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Opens"
          },
          "closes": {
            "type": "string",
            "pattern": "^([01]\\d|2[0-3]):[0-5]\\d$",
            "title": "Closes"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "weekday",
          "opens",
          "closes"
        ],
        "title": "ScheduleInput"
      },
      "UserInput": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 80,
            "minLength": 2,
            "title": "Name"
          },
          "email": {
            "type": "string",
            "maxLength": 120,
            "pattern": "^[^\\s@]+@[^\\s@]+\\.[^\\s@]+$",
            "title": "Email"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "UserInput"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}